🖱️ Combines deviceConfig and IMU service to peripherals
