+++ esp32/features.ini
@@ -1,18 +1,19 @@
 [features]
 build_flags = 
-  -D FT_BATTERY=0
+  -D FT_BATTERY=1
   -D FT_NTP=1
   -D FT_SECURITY=0
   -D FT_SLEEP=0
   -D FT_UPLOAD_FIRMWARE=0
   -D FT_DOWNLOAD_FIRMWARE=0
   -D FT_ANALYTICS=1
-  -D FT_MOTION=0
+  -D FT_MOTION=1
 
   ; Hardware specific
-  -D FT_IMU=0
-  -D FT_MAG=0
-  -D FT_BMP=0  
+  -D FT_IMU=1
+  -D FT_MAG=1
+  -D FT_BMP=1
   -D FT_GPS=0
-  -D FT_WS2812=0
+  -D FT_WS2812=1
   -D FT_SERVO=1
+  -D FT_ADS1115=1
